Update PHPUnit version and remove unused code

+++ config/Migrations/20240316165215_FixAreasData.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use Cake\ORM\Locator\LocatorAwareTrait;
 use Migrations\AbstractMigration;
+use Cake\I18n\FrozenTime;
 
 class FixAreasData extends AbstractMigration
 {
@@ -22,12 +23,14 @@ class FixAreasData extends AbstractMigration
         $programsTable = $this->fetchTable('Programs');
 
         foreach ($this->areas as $key => $areaData) {
-            $area = $areasTable->newEntity($areaData);
-            $area->created = date('Y-m-d H:i:s');
-            $area->modified = date('Y-m-d H:i:s');
+            if ($programsTable->hasField('areas') && $programsTable->exists(['area' => $key])) {
+                $area = $areasTable->newEntity($areaData);
+                $area->created = FrozenTime::now();
+                $area->modified = FrozenTime::now();
+                $area = $areasTable->saveOrFail($area);
 
-            $area = $areasTable->saveOrFail($area);
-            $programsTable->updateAll(['area_id' => $area->id], ['area' => $key]);
+                $programsTable->updateAll(['area_id' => $area->id], ['area' => $key]);
+            }
         }
 
         $table = $this->table('programs');
@@ -35,24 +38,27 @@ class FixAreasData extends AbstractMigration
         $table->update();
     }
 
-
     public function down(): void
     {
         $areasTable = $this->fetchTable('Areas');
         $programsTable = $this->fetchTable('Programs');
 
         $table = $this->table('programs');
-        $table->addColumn('area', 'string', [
-            'default' => null,
-            'limit' => 255,
-            'null' => false,
-        ]);
-        $table->update();
+        if (!$table->hasColumn('area')) {
+            $table->addColumn('area', 'string', [
+                'default' => null,
+                'limit' => 255,
+                'null' => false,
+            ]);
+            $table->update();
+        }
 
         foreach ($this->areas as $key => $areaData) {
             $area = $areasTable->find()->where(['abbr' => $areaData['abbr']])->first();
-            $programsTable->updateAll(['area' => $key], ['area_id' => $area->id]);
-            $areasTable->deleteOrFail($area);
+            if (!empty($area) && $programsTable->exists(['area_id' => $area->id])) {
+                $programsTable->updateAll(['area' => $key], ['area_id' => $area->id]);
+                $areasTable->deleteOrFail($area);
+            }
         }
     }
 }
